pyproject.toml
[project]
name = "f5-ai-gateway-sdk"
dynamic = ["version"]
authors = [
    { name = "Elijah Zupancic", email = "e.zupancic@f5.com" },
    { name = "Michael McAleer", email = "m.mcaleer@f5.com" },
    { name = "Eoin O'Shaughnessy", email = "e.oshaughnessy@f5.com" },
    { name = "Dan Lennon", email = "d.lennon@f5.com" },
    { name = "Andrei Zeliankou", email = "a.zelenkov@f5.com" },
]
description = "Python Starlette SDK for the F5 AI Gateway"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "starlette~=0.46.2",
    "python-multipart~=0.0.20",
    "requests-toolbelt~=1.0.0",
    "opentelemetry-sdk~=1.33.0",
    "opentelemetry-api~=1.33.0",
    "opentelemetry-semantic-conventions~=0.51b0",
    "pydantic~=2.11.4",
]

[project.urls]
Homepage = "https://github.com/nginxinc/f5-ai-gateway-sdk-py"
Issues = "https://github.com/nginxinc/f5-ai-gateway-sdk-py/issues"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
bump = true
strict = true
pattern = "default-unprefixed"

[tool.hatch.build.hooks.version]
path = "src/f5_ai_gateway_sdk/_version.py"
template = '''
version = "{version}"
'''

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
markers = ["test_of_tests", "wip_feature"]
minversion = "8.0"
pythonpath = "src"
log_cli = true
log_cli_level = "INFO"

[dependency-groups]
dev = [
    "httpx~=0.28.1",
    "pytest~=8.3.5",
    "pytest-cov~=6.1.1",
    "pytest-asyncio~=0.26.0",
    "parameterized>=0.9.0",
    "PyYAML>=6.0.2",
    "build>=1.2.2",
    "pip~=25.1.1",
    "setuptools~=80.4.0",
    "pyright==1.1.400",
    "bandit~=1.8.3",
    "ruff~=0.11.9",
]

[tool.pyright]
exclude = ["tests"]

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
# default rules and copyright check (CPY)
select = ["E4", "E7", "E9", "F", "CPY"]

[tool.ruff.lint.flake8-copyright]
author = "F5, Inc."
notice-rgx = "(?i)Copyright \\(C\\)"
